Extract additional facts
Moved fact gathering operations to a new file. Added rule for
gathering system profile facts for drift application.
